+++ src/main/java/au/com/sealink/quicktravel/client/QuickTravelApiClient.java
@@ -1,18 +1,6 @@
 package au.com.sealink.quicktravel.client;
 
-import au.com.sealink.quicktravel.client.models.BoardRequest;
-import au.com.sealink.quicktravel.client.models.BoardResult;
-import au.com.sealink.quicktravel.client.models.Booking;
-import au.com.sealink.quicktravel.client.models.Country;
-import au.com.sealink.quicktravel.client.models.EncryptedData;
-import au.com.sealink.quicktravel.client.models.PassengerType;
-import au.com.sealink.quicktravel.client.models.PaymentType;
-import au.com.sealink.quicktravel.client.models.ProductType;
-import au.com.sealink.quicktravel.client.models.Resource;
-import au.com.sealink.quicktravel.client.models.ResourceCategory;
-import au.com.sealink.quicktravel.client.models.Route;
-import au.com.sealink.quicktravel.client.models.Till;
-import au.com.sealink.quicktravel.client.models.User;
+import au.com.sealink.quicktravel.client.models.*;
 import au.com.sealink.quicktravel.client.models.checkout.Checkout;
 import au.com.sealink.quicktravel.client.models.checkout.CheckoutResponse;
 import au.com.sealink.quicktravel.client.models.reservationFor.core.Product;
@@ -120,4 +108,9 @@ public interface QuickTravelApiClient {
 
     @GET("/api/tills.json")
     Single<List<Till>> getTills();
+
+    @GET("/api/parties/clients.json")
+    Single<List<ClientSearchResponse>> searchClients(
+            @Query("search_query") String query
+    );
 }