+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * [TT-4450] Add network interceptor to change tenant hosts
 * [TT-4449] Add endpoint to create a new booking
 * [TT-4446] Add analytics support
+* [TT-4455] Add client search endpoint
 * [TT-4453] Expand booking object
 
 ## 0.5.0